refactor(cast): inline call implementation
Move RPC call handling into the call command and group its existing\nCLI coverage in a dedicated module.
